refactor: id long 타입으로 변경

+++ src/main/resources/data.sql
@@ -1,6 +1,6 @@
 -- TODO: 기능 구현에 필요한 내용을 추가하거나 수정하세요.
 CREATE TABLE PLAY_RESULT (
-    id          INT         NOT NULL AUTO_INCREMENT,
+    id          BIGINT         NOT NULL AUTO_INCREMENT,
     winners     VARCHAR(50) NOT NULL,
     trial_count INT NOT NULL,
     created_at  DATETIME    NOT NULL default current_timestamp,
@@ -8,10 +8,10 @@ CREATE TABLE PLAY_RESULT (
 );
 
 CREATE TABLE PLAYER (
-    id INT NOT NULL AUTO_INCREMENT,
-    play_result_id INT NOT NULL,
-    name VARCHAR(5) NOT NULL,
-    position INT NOT NULL,
+    id              BIGINT NOT NULL AUTO_INCREMENT,
+    play_result_id  INT NOT NULL,
+    name            VARCHAR(5) NOT NULL,
+    position        INT NOT NULL,
     PRIMARY KEY (id),
     FOREIGN KEY (play_result_id) REFERENCES PLAY_RESULT (id)
 );
